modified:   Groups/__pycache__/urls.cpython-313.pyc 	modified:   Groups/__pycache__/views.cpython-313.pyc 	modified:   Groups/urls.py 	modified:   Groups/views.py 	modified:   Post/__pycache__/urls.cpython-313.pyc 	modified:   Post/__pycache__/views.cpython-313.pyc 	modified:   Post/urls.py 	modified:   Post/views.py 	modified:   Profile/__pycache__/urls.cpython-313.pyc 	modified:   Profile/__pycache__/views.cpython-313.pyc 	modified:   Profile/urls.py 	modified:   Profile/views.py

+++ Groups/views.py
@@ -1,3 +1,4 @@
+from django.core.exceptions import ObjectDoesNotExist
 from django.contrib.auth.models import Group
 from rest_framework.response import Response
 from .serializers import GroupSerializer
@@ -11,28 +12,50 @@ class GroupAPIView(APIView):
     serializer_class = GroupSerializer
 
     def get(self,request, _pk):
-        group = Group.objects.get(pk=_pk)
-        if group.DoesNotExist:
-            return Response(status=status.HTTP_404_NOT_FOUND)
-        return Response(status=status.HTTP_200_OK,data=group.objects.all)
+        if request.user.is_authenticated():
+            try:
+                group = Group.objects.get(pk=_pk)
+                return Response(status=status.HTTP_200_OK,data=group.objects.all)
+            except ObjectDoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
     
     def post(self, request):
-        serializer = self.serializer_class(data = request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(status=status.HTTP_200_OK,data=serializer.data)
+        if request.user.is_authenticated():
+            try:
+                serializer = self.serializer_class(data = request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(status=status.HTTP_200_OK,data=serializer.data)
+                else:
+                    return Response(status=status.HTTP_400_BAD_REQUEST,data=serializer.data)
+            except ObjectDoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
         else:
-            return Response(status=status.HTTP_400_BAD_REQUEST,data=serializer.data)
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
     
     def patch(self, request,pk):
-        group = get_object_or_404(Group,pk=pk)
-        serializer = GroupSerializer(group, data= request.data, partial = True)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(status=status.HTTP_200_OK)
-        return Response(status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated():
+            try:
+                group = get_object_or_404(Group,pk=pk)
+                serializer = GroupSerializer(group, data= request.data, partial = True)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(status=status.HTTP_200_OK)
+                return Response(status=status.HTTP_400_BAD_REQUEST)
+            except ObjectDoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
     
     def delete(self, request, pk):
-        group = get_object_or_404(GroupSerializer,pk=pk)
-        group.delete()
-        return Response(status=status.HTTP_200_OK)
+        if request.user.is_authenticated():
+            try:
+                group = get_object_or_404(GroupSerializer,pk=pk)
+                group.delete()
+                return Response(status=status.HTTP_200_OK)
+            except ObjectDoesNotExist:
+                return Response(status=status.HTTP_404_NOT_FOUND)
+        else:
+            return Response(status=status.HTTP_401_UNAUTHORIZED)
